refactor: structure, naming
