+++ .prettierrc
@@ -1,4 +1,4 @@
 {
   "singleQuote": true,
-  "trailingComma": "all"
+  "trailingComma": "none"
 }
