+++ src/index.ts
@@ -1,4 +1,4 @@
-import { app } from './server'
+import { app } from './servers'
 import { getPortFromEnv, logServerStartedOnPort } from './helpers'
 
 const PORT = getPortFromEnv()
